feat: add dynamic routing for blog pages
